requirements.txt
Pillow==8.2.0
python_bcrypt==0.3.2
